[benc/tighten-alignment-parser] Tighten and parse the Alignment type
