+++ packages/perseus-core/src/data-schema.ts
@@ -316,6 +316,27 @@ export type PerseusImageDetail = {
     height: number;
 };
 
+// TODO(benchristel): as of 2026, only some of these alignments appear in our
+//  content corpus. Counts (for all locales):
+//     5955449  "block"
+//     20442080 "default"
+//     12622    "full-width"
+//     52       "wrap-left"
+//     35       "wrap-right"
+//  It seems possible that "wrap-left" and "wrap-right" are vestigial. "inline"
+//  and "inline-block" are definitely unused. Consider removing them from this
+//  type.
+export type Alignment =
+    | "default"
+    | "block"
+    | "inline-block"
+    | "inline"
+    // wrap alignments will be set to inline-block floated left or right this will
+    // allow text to wrap around the widget and not have large space on either side
+    | "wrap-left"
+    | "wrap-right"
+    | "full-width";
+
 /**
  * ItemExtras represent extra UI elements that help the learner in answering
  * the question (such as a calculator for questions where solving by hand is
@@ -409,7 +430,7 @@ export type WidgetOptions<
      * widget logic, which can be various other alignments (e.g.
      * "inline-block", "inline", etc).
      */
-    alignment?: string;
+    alignment?: Alignment;
     /**
      * Options specific to the type field of the widget. See Perseus*WidgetOptions for
      * more details
